+++ app/components/molecules/checkbox/termsCheckbox.tsx
@@ -1,16 +1,14 @@
 import React from 'react';
 import styles from './checkbox.module.css';
 
-interface CheckboxProps extends React.InputHTMLAttributes<HTMLInputElement> {
-  label?: string;
-}
 interface TermsCheckboxProps {
   onOpenModal: () => void;
   checked: boolean;
   onChange: (accepted: boolean) => void;
+  label?: string;
 }
 
-const TermsCheckbox: React.FC<TermsCheckboxProps> = ({ onOpenModal, checked, onChange }) => {
+const TermsCheckbox: React.FC<TermsCheckboxProps> = ({ onOpenModal, checked, onChange, label, ...props }) => {
   const handleCheckboxChange = () => {
     if (!checked) {
       onOpenModal(); 
@@ -18,10 +16,7 @@ const TermsCheckbox: React.FC<TermsCheckboxProps> = ({ onOpenModal, checked, onC
     onChange(!checked);
   };
 
-/**
- * Checkbox que funciona con React Hook Form, recibiendo onChange, value, etc. vía props.
- */
-const TermsCheckbox: React.FC<CheckboxProps> = ({ label, ...props }) => {
+
   return (
     <div className="mb-5">
       <label className={styles['checkbox-container']}>
@@ -37,6 +32,7 @@ const TermsCheckbox: React.FC<CheckboxProps> = ({ label, ...props }) => {
 
         {/* Si hay label, la renderizamos (puede ser "Acepto términos", etc.) */}
         {label && <span className="ml-2">{label}</span>}
+        Acepto los términos y condiciones de la plataforma y la política de privacidad.
       </label>
     </div>
   );
